# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for dev.",
        "title": "dtle API Docs",
        "contact": {},
        "version": "2.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v2/log/level": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "reload log level dynamically.",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "tags": [
                    "log"
                ],
                "operationId": "UpdateLogLevelV2",
                "parameters": [
                    {
                        "enum": [
                            "TRACE",
                            "DEBUG",
                            "INFO",
                            "WARN",
                            "ERROR"
                        ],
                        "type": "string",
                        "description": "dtle log level",
                        "name": "dtle_log_level",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UpdataLogLevelRespV2"
                        }
                    }
                }
            }
        },
...
        "/v2/monitor/task": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "get progress of tasks within an allocation.",
                "tags": [
                    "monitor"
                ],
                "operationId": "GetTaskProgressV2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "allocation id",
                        "name": "allocation_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "task name",
                        "name": "task_name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "nomad_http_address is the http address of the nomad that the target dtle is running with. ignore it if you are not sure what to provide",
                        "name": "nomad_http_address",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.GetTaskProgressRespV2"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.BufferStat": {
            "type": "object",
            "properties": {
                "applier_tx_queue_size": {
                    "type": "integer"
                },
                "binlog_event_queue_size": {
                    "type": "integer"
                },
                "extractor_tx_queue_size": {
                    "type": "integer"
                },
                "send_by_size_full": {
                    "type": "integer"
                },
                "send_by_timeout": {
                    "type": "integer"
                }
            }
        },
...
        "models.CurrentCoordinates": {
            "type": "object",
            "properties": {
                "file": {
                    "type": "string"
                },
                "gtid_set": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "read_master_log_pos": {
                    "type": "integer"
                },
                "relay_master_log_file": {
                    "type": "string"
                },
                "retrieved_gtid_set": {
                    "type": "string"
                }
            }
        },
...
        "models.DelayCount": {
            "type": "object",
            "properties": {
                "num": {
                    "type": "integer"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
...
        "models.GetTaskProgressRespV2": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "tasks_status": {
                    "$ref": "#/definitions/models.TaskProgress"
                }
            }
        },
...
        "models.NatsMessageStatistics": {
            "type": "object",
            "properties": {
                "in_bytes": {
                    "type": "integer"
                },
                "in_messages": {
                    "type": "integer"
                },
                "out_bytes": {
                    "type": "integer"
                },
                "out_messages": {
                    "type": "integer"
                },
                "reconnects": {
                    "type": "integer"
                }
            }
        },
...
        "models.TaskProgress": {
            "type": "object",
            "properties": {
                "ETA": {
                    "type": "string"
                },
                "backlog": {
                    "type": "string"
                },
                "buffer_status": {
                    "$ref": "#/definitions/models.BufferStat"
                },
                "current_coordinates": {
                    "$ref": "#/definitions/models.CurrentCoordinates"
                },
                "delay_count": {
                    "$ref": "#/definitions/models.DelayCount"
                },
                "exec_master_row_count": {
                    "type": "integer"
                },
                "exec_master_tx_count": {
                    "type": "integer"
                },
                "nats_message_status": {
                    "$ref": "#/definitions/models.NatsMessageStatistics"
                },
                "progress_PCT": {
                    "type": "string"
                },
                "read_master_row_count": {
                    "type": "integer"
                },
                "read_master_tx_count": {
                    "type": "integer"
                },
                "stage": {
                    "type": "string"
                },
                "throughput_status": {
                    "$ref": "#/definitions/models.ThroughputStat"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
...
        "models.ThroughputStat": {
            "type": "object",
            "properties": {
                "num": {
                    "type": "integer"
                },
                "time": {
                    "type": "integer"
                }
            }
        },
        "models.UpdataLogLevelRespV2": {
            "type": "object",
            "properties": {
                "dtle_log_level": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}